updated eb environment config, cast vars to string where necessary to avoid bugs,  moved save_video and upload_file_to_s3 functions into helper.py
