+++ .ebextensions/01_environment.config
@@ -7,11 +7,6 @@
       "value" : "Immutable"
       },
       {
-      "namespace" : "aws:elasticbeanstalk:command",
-      "option_name" : "DeploymentPolicy",
-      "value" : "Immutable"
-      },
-      {
       "namespace" : "aws:ec2:instances",
       "option_name" : "InstanceTypes",
       "value" : "p3.2xlarge"
